KAS-1407: refactor item-document
